Move format specifier lexing from syntax to ide_db
